Refactor prosody parser code structure
