+++ .kilo/plans/1776591113940-sunny-falcon.md
@@ -1,0 +1,75 @@
+# Plan: Code Structure Cleanup for thepulimaangani-parser
+
+## Goal
+Refactor the monolithic `lib.rs` file into modular Rust code following the CODING_STYLE.md guidelines, improving maintainability and readability.
+
+## Current State
+- All code is in a single 1549-line `lib.rs` file
+- Contains types, parsing logic, WASM bindings, and tests
+- Working prototype exists with tests passing
+
+## Prerequisites
+- Tests are already implemented and passing (✓)
+- CODING_STYLE.md provides clear guidelines for structure
+
+## Next Steps
+
+### 1. Create Module Structure
+- Create `types.rs` for core data structures (SyllableType, Syllable, Foot, Line, etc.)
+- Create `chars.rs` for Tamil character sets and constants
+- Create `parse.rs` for parsing functions (detect_syllables, parse_lines, etc.)
+- Create `metre.rs` for metre detection logic (check_venpaa, check_kaliviruttam, etc.)
+- Create `talai.rs` for bond/talai calculation functions
+- Create `utils.rs` for utility functions (remove_punctuation, tamil_to_romanized, etc.)
+
+### 2. Move Constants to chars.rs
+- SYLLABLE_TYPES, WORD_TYPES, VENPAA_WORD_CLASS
+- Tamil character arrays (independent_vowels, vowel_signs, consonants, etc.)
+- get_word_type function
+
+### 3. Move Types to types.rs
+- SyllableType enum
+- Syllable, Foot, Line, LetterCount, ParseResult structs
+
+### 4. Move Parsing Logic to parse.rs
+- get_letter_count
+- detect_syllables
+- parse_lines
+- parse_poem (as entry point, but delegate to modules)
+
+### 5. Move Metre Detection to metre.rs
+- check_venpaa, check_venpaavinam, check_kaliviruttam, etc.
+- get_metre_type
+
+### 6. Move Talai Logic to talai.rs
+- calculate_talai
+- get_bond_type
+
+### 7. Move Utilities to utils.rs
+- remove_punctuation
+- tamil_to_romanized
+
+### 8. Update lib.rs
+- Keep only WASM bindings and module declarations
+- Export public functions
+- Tests can stay or be moved to separate test modules
+
+### 9. Update Cargo.toml
+- Ensure dependencies are minimal (serde, wasm-bindgen, regex)
+- Add any needed dev-dependencies
+
+### 10. Run Tests and Lint
+- Run `cargo test` to ensure functionality preserved
+- Run `cargo clippy --all-targets -- -D warnings` for style compliance
+- Run `cargo fmt` for formatting
+
+## Benefits
+- Improved code organization and maintainability
+- Easier to locate and modify specific functionality
+- Better adherence to Rust best practices
+- Reduced cognitive load when working on individual features
+
+## Risk Mitigation
+- Run tests after each major move to ensure no regressions
+- Keep WASM bindings as the last step to maintain working prototype
+- Use Rust's module system properly with `pub mod` declarations
