+++ rust-parser/src/lib.rs
@@ -1,249 +1,32 @@
-use regex::Regex;
-use serde::{Deserialize, Serialize};
-use std::collections::HashMap;
 use wasm_bindgen::prelude::*;
 
-// ==================== CORE TYPES (mirrors PHP ProsodyParseTree) ====================
+mod chars;
+mod metre;
+mod parse;
+mod talai;
+mod types;
+mod utils;
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub enum SyllableType {
-    Ner,   // நேர்
-    Nirai, // நிரை
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Syllable {
-    pub text: String,
-    pub syllable_type: SyllableType,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Foot {
-    pub syllables: Vec<Syllable>,
-    pub foot_type: String, // e.g. "tEmA", "puLimA", "mA", etc. (from original WordType)
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct Line {
-    pub feet: Vec<Foot>,
-    pub line_class: String, // kuRaLaTi, ci_ntaTi, etc.
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct LetterCount {
-    pub vowel: usize,
-    pub consonant: usize,
-    pub consonant_vowel: usize,
-    pub aytham: usize,
-    pub short: usize,
-    pub long: usize,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct ParseResult {
-    pub original_text: String,
-    pub lines: Vec<Line>,
-    pub metre_type: String,
-    pub letter_count: LetterCount,
-    pub vikalpa_count: usize,
-    pub word_bond: String, // talai linkages (will expand later)
-    pub errors: Vec<String>,
-}
+use crate::chars::get_word_type;
+use crate::metre::{
+    check_asiriyappaa, check_kalippaa, check_venkalippaa, check_venpaa, check_venpaa_multiline,
+    get_metre_type,
+};
+use crate::parse::*;
+use crate::talai::{calculate_talai, get_bond_type};
+use crate::types::*;
+use crate::utils::{remove_punctuation, tamil_to_romanized};
 
 // ==================== REFERENCE DATA (ported directly from PHP) ====================
 
-const SYLLABLE_TYPES: [&str; 2] = ["nE_r", "nirY"];
-
-const WORD_TYPES: [(&str, &str); 6] = [
-    ("nE_rnE_r", "tEmA"),
-    ("nirYnE_r", "puLimA"),
-    ("nE_rnirY", "kUviLa_m"),
-    ("nirYnirY", "karuviLa_m"),
-    // ... (full list from original PHP — I included the most common ones; we’ll expand as needed)
-    ("nE_r", "mA"),
-    ("nirY", "viLa_m"),
-    // Add the rest of the 3-asai, 4-asai, etc. from the original WordType array here
-];
-
-const VENPAA_WORD_CLASS: [(&str, &str); 4] = [
-    ("nE_r", "nA_L"),
-    ("nirY", "mala_r"),
-    ("nE_rpu", "kAcu"),
-    ("nirYpu", "piRa_ppu"),
-];
-
-// Foot classification map (WordType from PHP) - expanded with more traditional types
-fn get_word_type(pattern: &str) -> &'static str {
-    match pattern {
-        // Basic 2-syllable feet
-        "nE_rnE_r" => "tEmA",
-        "nirYnE_r" => "puLimA",
-        "nE_rnirY" => "kUviLa_m",
-        "nirYnirY" => "karuviLa_m",
-
-        // 3-syllable feet (angaay)
-        "nE_rnE_rnE_r" => "tEmA_GkA_y",
-        "nirYnE_rnE_r" => "puLimA_GkA_y",
-        "nE_rnirYnE_r" => "kUviLa_GkA_y",
-        "nirYnirYnE_r" => "karuviLa_GkA_y",
-
-        // 3-syllable feet (kavi)
-        "nE_rnE_rnirY" => "tEmA_GkaVi",
-        "nirYnE_rnirY" => "puLimA_GkaVi",
-        "nE_rnirYnirY" => "kUviLa_GkaVi",
-        "nirYnirYnirY" => "karuviLa_GkaVi",
-
-        // 4-syllable feet (puu)
-        "nE_rnE_rnE_rnE_r" => "tEmA_nta_NpU",
-        "nirYnE_rnE_rnE_r" => "puLimA_nta_NpU",
-        "nE_rnirYnE_rnE_r" => "kUviLa_nta_NpU",
-        "nirYnirYnE_rnE_r" => "karuviLa_nta_NpU",
-
-        // 4-syllable feet (arum puu)
-        "nE_rnE_rnirYnE_r" => "tEmAnaRu_mpU",
-        "nirYnE_rnirYnE_r" => "puLimAnaRu_mpU",
-        "nE_rnirYnirYnE_r" => "kUviLanaRu_mpU",
-        "nirYnirYnirYnE_r" => "karuviLanaRu_mpU",
-
-        // 4-syllable feet (nizhal)
-        "nE_rnE_rnirYnirY" => "tEmAnaRuniZa_l",
-        "nirYnE_rnirYnirY" => "puLimAnaRuniZa_l",
-        "nE_rnirYnirYnirY" => "kUviLanaRuniZa_l",
-        "nirYnirYnirYnirY" => "karuviLanaRuniZa_l",
-
-        // 4-syllable feet (nta nizhal)
-        "nE_rnE_rnE_rnirY" => "tEmA_nta_NNiZa_l",
-        "nirYnE_rnE_rnirY" => "puLimA_nta_NNiZa_l",
-        "nE_rnirYnE_rnirY" => "kUviLa_nta_NNiZa_l",
-        "nirYnirYnE_rnirY" => "karuviLa_nta_NNiZa_l",
-
-        // Single syllable feet
-        "nE_r" => "mA",
-        "nirY" => "viLa_m",
-
-        // 5-syllable feet
-        "nE_rnE_rnE_rnE_rnE_r" => "tEmA_GkA_yi_Ra_pU",
-        "nirYnE_rnE_rnE_rnE_r" => "puLimA_GkA_yi_Ra_pU",
-        "nE_rnirYnE_rnE_rnE_r" => "kUviLa_GkA_yi_Ra_pU",
-        "nirYnirYnE_rnE_rnE_r" => "karuviLa_GkA_yi_Ra_pU",
-
-        _ => "unknown",
-    }
-}
-
 // ==================== WASM ENTRY POINT ====================
-
-fn get_letter_count(text: &str) -> LetterCount {
-    let tamil_text = text.trim();
-    let chars: Vec<char> = tamil_text.chars().collect();
-
-    // Tamil character sets
-    let independent_vowels = ["அ", "ஆ", "இ", "ஈ", "உ", "ஊ", "எ", "ஏ", "ஐ", "ஒ", "ஓ", "ஔ"];
-    let vowel_signs = ["ா", "ி", "ீ", "ு", "ூ", "ெ", "ே", "ை", "ொ", "ோ", "ௌ"];
-    let consonants = [
-        "க", "ங", "ச", "ஜ", "ஞ", "ட", "ண", "த", "ந", "ன", "ப", "ம", "ய", "ர", "ற", "ல", "ள", "ழ",
-        "வ", "ஶ", "ஷ", "ஸ", "ஹ",
-    ];
-    let virama = "்";
-    let aytham = "ஃ";
-
-    let mut vowel_count = 0;
-    let mut consonant_count = 0; // pure consonants (with virama)
-    let mut consonant_vowel_count = 0; // consonants with vowels
-    let mut aytham_count = 0;
-    let mut short_vowel_count = 0;
-    let mut long_vowel_count = 0;
-
-    let mut i = 0;
-    while i < chars.len() {
-        let current = chars[i].to_string();
-
-        // Count aytham
-        if current == aytham {
-            aytham_count += 1;
-            i += 1;
-            continue;
-        }
-
-        // Count independent vowels
-        if independent_vowels.contains(&current.as_str()) {
-            vowel_count += 1;
-
-            // Classify short vs long vowels
-            match current.as_str() {
-                "அ" | "இ" | "உ" | "எ" | "ஒ" => short_vowel_count += 1,
-                _ => long_vowel_count += 1,
-            }
-
-            i += 1;
-            continue;
-        }
-
-        // Handle consonants
-        if consonants.contains(&current.as_str()) {
-            i += 1;
-
-            // Check what follows the consonant
-            if i < chars.len() {
-                let next = chars[i].to_string();
-
-                if next == virama {
-                    // Pure consonant (consonant + virama)
-                    consonant_count += 1;
-                    i += 1; // skip virama
-                } else if vowel_signs.contains(&next.as_str()) {
-                    // Consonant + vowel sign = consonant-vowel combination
-                    consonant_vowel_count += 1;
-                    vowel_count += 1;
-
-                    // Classify vowel sign as short or long
-                    match next.as_str() {
-                        "ி" | "ு" | "ெ" | "ொ" => short_vowel_count += 1,
-                        _ => long_vowel_count += 1,
-                    }
-
-                    i += 1; // skip vowel sign
-
-                    // Check for aytham after vowel sign
-                    if i < chars.len() && chars[i].to_string() == aytham {
-                        aytham_count += 1;
-                        i += 1;
-                    }
-                } else {
-                    // Bare consonant (shouldn't happen in proper Tamil, but count as consonant-vowel with implicit 'a')
-                    consonant_vowel_count += 1;
-                    vowel_count += 1;
-                    short_vowel_count += 1; // implicit short 'a'
-                }
-            } else {
-                // Consonant at end of text (implicit 'a')
-                consonant_vowel_count += 1;
-                vowel_count += 1;
-                short_vowel_count += 1; // implicit short 'a'
-            }
-            continue;
-        }
-
-        // Skip other characters (spaces, punctuation, etc.)
-        i += 1;
-    }
-
-    LetterCount {
-        vowel: vowel_count,
-        consonant: consonant_count,
-        consonant_vowel: consonant_vowel_count,
-        aytham: aytham_count,
-        short: short_vowel_count,
-        long: long_vowel_count,
-    }
-}
 
 #[wasm_bindgen]
 pub fn parse_poem(tamil_text: &str) -> String {
     // Basic preprocessing
     let cleaned = tamil_text.trim().to_string();
 
-    let letter_count_struct = get_letter_count(&cleaned);
+    let letter_count_struct: types::LetterCount = get_letter_count(&cleaned);
 
     // Parse lines and feet
     let lines = parse_lines(&cleaned);
@@ -258,7 +41,7 @@ pub fn parse_poem(tamil_text: &str) -> String {
     // Determine metre type
     let metre_type = get_metre_type(&lines, total_bonds, kali_bonds, ven_bonds);
 
-    let result = ParseResult {
+    let result = types::ParseResult {
         original_text: cleaned.clone(),
         lines,
         metre_type,
@@ -272,637 +55,12 @@ pub fn parse_poem(tamil_text: &str) -> String {
 }
 
 // Tamil to Romanized transliteration (based on PHP tam2lat)
-fn tamil_to_romanized(text: &str) -> String {
-    let mut tameng: HashMap<&str, &str> = HashMap::new();
-    tameng.insert("அ", "_a");
-    tameng.insert("ஆ", "_A");
-    tameng.insert("இ", "_i");
-    tameng.insert("ஈ", "_I");
-    tameng.insert("உ", "_u");
-    tameng.insert("ஊ", "_U");
-    tameng.insert("எ", "_e");
-    tameng.insert("ஏ", "_E");
-    tameng.insert("ஐ", "_Y");
-    tameng.insert("ஒ", "_o");
-    tameng.insert("ஓ", "_O");
-    tameng.insert("ஔ", "_W");
-    tameng.insert("க", "k");
-    tameng.insert("ங", "G");
-    tameng.insert("ச", "c");
-    tameng.insert("ஜ", "j");
-    tameng.insert("ஞ", "J");
-    tameng.insert("ட", "T");
-    tameng.insert("ண", "N");
-    tameng.insert("த", "t");
-    tameng.insert("ந", "n");
-    tameng.insert("ன", "V");
-    tameng.insert("ப", "p");
-    tameng.insert("ம", "m");
-    tameng.insert("ய", "y");
-    tameng.insert("ர", "r");
-    tameng.insert("ற", "R");
-    tameng.insert("ல", "l");
-    tameng.insert("ள", "L");
-    tameng.insert("ழ", "Z");
-    tameng.insert("வ", "v");
-    tameng.insert("ஶ", "F");
-    tameng.insert("ஷ", "S");
-    tameng.insert("ஸ", "s");
-    tameng.insert("ஹ", "h");
-    tameng.insert("ா", "A");
-    tameng.insert("ி", "i");
-    tameng.insert("ீ", "I");
-    tameng.insert("ு", "u");
-    tameng.insert("ூ", "U");
-    tameng.insert("ெ", "e");
-    tameng.insert("ே", "E");
-    tameng.insert("ை", "Y");
-    tameng.insert("ொ", "o");
-    tameng.insert("ோ", "O");
-    tameng.insert("ௌ", "W");
-    tameng.insert("ஃ", "_K");
-
-    let mut result = text.to_string();
-
-    // Replace consonant + virama with consonant
-    let re = Regex::new(r"_([kGcJTNtnpmyrlvZLRVjSsh])").unwrap();
-    result = re.replace_all(&result, "$1்").to_string();
-
-    // Replace characters
-    for (tamil, roman) in &tameng {
-        result = result.replace(tamil, roman);
-    }
-
-    // Special replacements
-    result = result.replace("a", "").replace("B", "ௌ").replace("Q", "ை");
-    result = result.replace("_", "");
-
-    result
-}
-
-// Improved syllable detection following Tamil phonological rules
-fn detect_syllables(text: &str) -> Vec<(String, SyllableType)> {
-    let mut syllables = Vec::new();
-
-    // Tamil character sets
-    let independent_vowels = ["அ", "ஆ", "இ", "ஈ", "உ", "ஊ", "எ", "ஏ", "ஐ", "ஒ", "ஓ", "ஔ"];
-    let short_vowel_signs = ["ி", "ு", "ெ", "ொ"];
-    let long_vowel_signs = ["ா", "ீ", "ூ", "ே", "ோ", "ௌ", "ை"];
-    let consonants = [
-        "க", "ங", "ச", "ஜ", "ஞ", "ட", "ண", "த", "ந", "ன", "ப", "ம", "ய", "ர", "ற", "ல", "ள", "ழ",
-        "வ", "ஶ", "ஷ", "ஸ", "ஹ",
-    ];
-    let virama = "்";
-    let aytham = "ஃ";
-
-    let chars: Vec<char> = text.chars().collect();
-    let mut i = 0;
-
-    while i < chars.len() {
-        let mut syllable = String::new();
-
-        // Skip whitespace and punctuation
-        if chars[i].is_whitespace() || !chars[i].is_alphabetic() {
-            i += 1;
-            continue;
-        }
-
-        let current = chars[i].to_string();
-
-        // Case 1: Independent vowel
-        if independent_vowels.contains(&current.as_str()) {
-            syllable.push(chars[i]);
-            i += 1;
-
-            // Check for aytham
-            if i < chars.len() && chars[i].to_string() == aytham {
-                syllable.push(chars[i]);
-                i += 1;
-            }
-
-            // Independent vowels are Ner (short) except for long ones
-            let is_long = matches!(current.as_str(), "ஆ" | "ஈ" | "ஊ" | "ஏ" | "ஐ" | "ஓ" | "ஔ");
-            let syllable_type = if is_long {
-                SyllableType::Nirai
-            } else {
-                SyllableType::Ner
-            };
-            syllables.push((syllable, syllable_type));
-            continue;
-        }
-
-        // Case 2: Consonant-based syllable
-        if consonants.contains(&current.as_str()) {
-            syllable.push(chars[i]);
-            i += 1;
-
-            // Check what follows
-            if i < chars.len() {
-                let next = chars[i].to_string();
-
-                // Case 2a: Consonant + vowel sign
-                if short_vowel_signs.contains(&next.as_str())
-                    || long_vowel_signs.contains(&next.as_str())
-                {
-                    syllable.push(chars[i]);
-                    i += 1;
-
-                    // Check for aytham after vowel sign
-                    if i < chars.len() && chars[i].to_string() == aytham {
-                        syllable.push(chars[i]);
-                        i += 1;
-                    }
-
-                    // Determine type based on vowel sign
-                    let syllable_type = if long_vowel_signs.contains(&next.as_str()) {
-                        SyllableType::Nirai
-                    } else {
-                        SyllableType::Ner
-                    };
-                    syllables.push((syllable, syllable_type));
-                    continue;
-                }
-
-                // Case 2b: Consonant + virama (pure consonant)
-                if next == virama {
-                    syllable.push(chars[i]);
-                    i += 1;
-
-                    // Pure consonants are Ner
-                    syllables.push((syllable, SyllableType::Ner));
-                    continue;
-                }
-
-                // Case 2c: Consonant + consonant (implicit vowel, but should be handled by next iteration)
-                // For now, treat as consonant with implicit short vowel
-                syllables.push((syllable, SyllableType::Ner));
-                continue;
-            } else {
-                // Consonant at end - implicit short vowel
-                syllables.push((syllable, SyllableType::Ner));
-                continue;
-            }
-        }
-
-        // Skip unknown characters
-        i += 1;
-    }
-
-    syllables
-}
 
 // Basic metre validation (simplified CheckVenpaa)
-fn check_venpaa(feet: &[Foot]) -> bool {
-    // Venpaa typically has 4 feet in first line, 3 in last line
-    // For now, just check if we have feet and they are of allowed types
-    let allowed_types = [
-        "tEmA",
-        "puLimA",
-        "kUviLa_m",
-        "karuviLa_m",
-        "tEmA_GkA_y",
-        "puLimA_GkA_y",
-        "kUviLa_GkA_y",
-        "karuviLa_GkA_y",
-    ];
-
-    for foot in feet {
-        if !allowed_types.contains(&foot.foot_type.as_str()) {
-            return false;
-        }
-    }
-
-    true
-}
-
-// Parse lines from text
-fn parse_lines(text: &str) -> Vec<Line> {
-    let mut lines = Vec::new();
-
-    for line_text in text.split('\n') {
-        let cleaned_line = remove_punctuation(line_text.trim());
-        if cleaned_line.is_empty() {
-            continue;
-        }
-
-        let mut feet = Vec::new();
-
-        for word in cleaned_line.split_whitespace() {
-            let detected_syllables = detect_syllables(word);
-            if detected_syllables.is_empty() {
-                continue;
-            }
-
-            let syllables: Vec<Syllable> = detected_syllables
-                .into_iter()
-                .map(|(text, syl_type)| Syllable {
-                    text,
-                    syllable_type: syl_type,
-                })
-                .collect();
-
-            let pattern = syllables
-                .iter()
-                .map(|s| match s.syllable_type {
-                    SyllableType::Ner => "nE_r",
-                    SyllableType::Nirai => "nirY",
-                })
-                .collect::<Vec<_>>()
-                .join("");
-
-            let foot_type = get_word_type(&pattern);
-
-            let foot = Foot {
-                syllables,
-                foot_type: foot_type.to_string(),
-            };
-
-            feet.push(foot);
-        }
-
-        let line = Line {
-            feet,
-            line_class: "kuRaLaTi".to_string(), // TODO: determine line class
-        };
-
-        lines.push(line);
-    }
-
-    lines
-}
-
-// Remove punctuation from text (simplified)
-fn remove_punctuation(text: &str) -> String {
-    text.chars()
-        .filter(|c| !c.is_ascii_punctuation() && *c != '।' && *c != '॥') // Add Tamil punctuation if needed
-        .collect()
-}
 
 // Calculate talai (bonds/linkages) between feet
-fn calculate_talai(lines: &[Line]) -> (usize, usize, usize) {
-    let mut total_bonds = 0;
-    let mut kali_bonds = 0;
-    let mut ven_bonds = 0;
-
-    for line in lines {
-        for window in line.feet.windows(2) {
-            total_bonds += 1;
-            let prev_foot = &window[0];
-            let curr_foot = &window[1];
-
-            let bond_type = get_bond_type(
-                &prev_foot.foot_type,
-                curr_foot
-                    .syllables
-                    .first()
-                    .map(|s| s.syllable_type.clone())
-                    .unwrap_or(SyllableType::Ner),
-            );
-
-            if bond_type.contains("கலித்தளை") {
-                kali_bonds += 1;
-            }
-            if bond_type.contains("வெண்டளை") {
-                ven_bonds += 1;
-            }
-        }
-    }
-
-    (total_bonds, kali_bonds, ven_bonds)
-}
-
-// Get bond type between two feet
-fn get_bond_type(prev_foot_type: &str, next_syllable_type: SyllableType) -> String {
-    let next_syl = match next_syllable_type {
-        SyllableType::Ner => "nE_r",
-        SyllableType::Nirai => "nirY",
-    };
-
-    // Based on PHP logic, using contains instead of ends_with for kA_y and pU
-    if prev_foot_type.contains("mA") && next_syl == "nE_r" {
-        "நேரொன்றிய ஆசிரியத்தளை".to_string()
-    } else if prev_foot_type.contains("viLa_m") && next_syl == "nirY" {
-        "நிரையொன்றிய ஆசிரியத்தளை".to_string()
-    } else if (prev_foot_type.contains("mA") && next_syl == "nirY")
-        || (prev_foot_type.contains("viLa_m") && next_syl == "nE_r")
-    {
-        "இயற்சீர் வெண்டளை".to_string()
-    } else if (prev_foot_type.contains("pU") || prev_foot_type.contains("kA_y"))
-        && next_syl == "nE_r"
-    {
-        "வெண்சீர் வெண்டளை".to_string()
-    } else if (prev_foot_type.contains("NiZa_l")
-        || prev_foot_type.contains("niZa_l")
-        || prev_foot_type.contains("kaVi"))
-        && next_syl == "nirY"
-    {
-        "ஒன்றிய வஞ்சித்தளை".to_string()
-    } else if (prev_foot_type.contains("NiZa_l")
-        || prev_foot_type.contains("niZa_l")
-        || prev_foot_type.contains("kaVi"))
-        && next_syl == "nE_r"
-    {
-        "ஒன்றா வஞ்சித்தளை".to_string()
-    } else if (prev_foot_type.contains("pU") || prev_foot_type.contains("kA_y"))
-        && next_syl == "nirY"
-    {
-        "கலித்தளை".to_string()
-    } else {
-        "unknown".to_string()
-    }
-}
 
 // Get metre type by checking in priority order
-fn get_metre_type(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> String {
-    if let Some(metre) = check_venpaa_multiline(lines) {
-        return metre;
-    }
-    if let Some(metre) = check_venpaavinam(lines, total_bonds, kali_bonds, ven_bonds) {
-        return metre;
-    }
-    if let Some(metre) = check_kaliviruttam(lines, total_bonds, kali_bonds, ven_bonds) {
-        return metre;
-    }
-    if let Some(metre) = check_asiriyappaa(lines, total_bonds, kali_bonds, ven_bonds) {
-        return metre;
-    }
-    if let Some(metre) = check_kalippaa(lines, total_bonds, kali_bonds, ven_bonds) {
-        return metre;
-    }
-    if let Some(metre) = check_venkalippaa(lines, total_bonds, kali_bonds, ven_bonds) {
-        return metre;
-    }
-    // TODO: add more checks like Vanjippaa, etc.
-    "Unknown".to_string()
-}
-
-// Check if Venpaa (simplified)
-fn check_venpaa_multiline(lines: &[Line]) -> Option<String> {
-    if lines.len() != 1 {
-        return None;
-    }
-    let line = &lines[0];
-    if check_venpaa(&line.feet) {
-        Some("வெண்பா (Venpaa)".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if multi-line Venbaa (2-line Venbaa)
-fn check_venpaavinam(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> Option<String> {
-    // Multi-line Venbaa typically has 2 lines with 4 + 3 feet pattern
-    if lines.len() != 2 {
-        return None;
-    }
-
-    let foot_counts: Vec<usize> = lines.iter().map(|l| l.feet.len()).collect();
-    if foot_counts != [4, 3] {
-        return None;
-    }
-
-    // For Venpaavinam, we allow some "unknown" feet as they might be complex valid patterns
-    // that our current foot type mapping doesn't cover yet
-    let unknown_count = lines
-        .iter()
-        .flat_map(|line| &line.feet)
-        .filter(|foot| foot.foot_type == "unknown")
-        .count();
-
-    // Allow up to 2 unknown feet (to be more permissive during development)
-    if unknown_count > 2 {
-        return None;
-    }
-
-    // Basic bonding check
-    if total_bonds > 0 {
-        Some("வெண்பா (Venpaa)".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if Kaliviruttam (கலிவிருத்தம்)
-fn check_kaliviruttam(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> Option<String> {
-    // Kaliviruttam typically has 4 lines with 4 feet each (4×4 pattern)
-    if lines.len() != 4 {
-        return None;
-    }
-
-    // Each line must have exactly 4 feet
-    for line in lines {
-        if line.feet.len() != 4 {
-            return None;
-        }
-    }
-
-    // Allow some "unknown" feet for complex patterns (up to 6 total for kaliviruttam)
-    let unknown_count = lines
-        .iter()
-        .flat_map(|line| &line.feet)
-        .filter(|foot| foot.foot_type == "unknown")
-        .count();
-
-    if unknown_count > 6 {
-        return None;
-    }
-
-    // Kaliviruttam requires some bonding but is more flexible than asiriyappaa
-
-    if total_bonds > 0 {
-        Some("kaliviru_tta_m".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if Kaliviruttam (கலிவிருத்தம்)
-fn check_kaliviruttam(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> Option<String> {
-    // Kaliviruttam typically has 4 lines with 4 feet each (4×4 pattern)
-    if lines.len() != 4 {
-        return None;
-    }
-
-    // Each line must have exactly 4 feet
-    for line in lines {
-        if line.feet.len() != 4 {
-            return None;
-        }
-    }
-
-    // Allow some "unknown" feet for complex patterns (up to 6 total for kaliviruttam)
-    let unknown_count = lines
-        .iter()
-        .flat_map(|line| &line.feet)
-        .filter(|foot| foot.foot_type == "unknown")
-        .count();
-
-    if unknown_count > 6 {
-        return None;
-    }
-
-    // Kaliviruttam requires some bonding but is more flexible than asiriyappaa
-
-    if total_bonds > 0 {
-        Some("kaliviru_tta_m".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if Venkalippaa
-fn check_venkalippaa(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> Option<String> {
-    // Must have more than 3 lines
-    if lines.len() <= 3 {
-        return None;
-    }
-
-    // Check foot counts per line
-    let mut line_class_check = true;
-    let final_line_index = lines.len() - 1;
-    for (i, line) in lines.iter().enumerate() {
-        if i < final_line_index {
-            // Non-final lines must have exactly 4 feet
-            if line.feet.len() != 4 {
-                line_class_check = false;
-            }
-        } else {
-            // Final line must have exactly 3 feet
-            if line.feet.len() != 3 {
-                return None;
-            }
-        }
-    }
-
-    // Check disallowed foot types (no tEmA, puLimA, karuviLa_GkaVi, kUviLa_GkaVi) only in non-final lines
-    let disallowed = ["tEmA", "puLimA", "karuviLa_GkaVi", "kUviLa_GkaVi"];
-    let mut word_class_check = true;
-    for (i, line) in lines.iter().enumerate() {
-        if i < final_line_index {
-            for foot in &line.feet {
-                if disallowed.contains(&foot.foot_type.as_str()) {
-                    word_class_check = false;
-                }
-            }
-        }
-    }
-
-    // Talai check: ≥50% kali+ven bonds, ≥25% kali bonds
-    let talai_check = if total_bonds > 0 {
-        let combined = kali_bonds + ven_bonds;
-        (combined as f64 / total_bonds as f64) > 0.5
-            && (kali_bonds as f64 / total_bonds as f64) > 0.25
-    } else {
-        false
-    };
-
-    if line_class_check && word_class_check && talai_check {
-        Some("ve_Nkali_ppA".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if Asiriyappaa
-fn check_asiriyappaa(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    ven_bonds: usize,
-) -> Option<String> {
-    // Asiriyappaa typically has 4 lines, but can be analyzed per 4-line stanza
-    // For now, check if we have exactly 4 lines
-    if lines.len() != 4 {
-        return None;
-    }
-
-    // Each line must have exactly 4 feet
-    for line in lines {
-        if line.feet.len() != 4 {
-            return None;
-        }
-    }
-
-    // Talai check: Asiriyappaa requires strict bonding
-    // Typically ≥60% bonded feet with good kali/ven distribution
-    let talai_check = if total_bonds > 0 {
-        let combined = kali_bonds + ven_bonds;
-        (combined as f64 / total_bonds as f64) > 0.6 && kali_bonds >= ven_bonds // More kali than ven bonds
-    } else {
-        false
-    };
-
-    if talai_check {
-        Some("Aciriya_ppA".to_string())
-    } else {
-        None
-    }
-}
-
-// Check if Kalippaa
-fn check_kalippaa(
-    lines: &[Line],
-    total_bonds: usize,
-    kali_bonds: usize,
-    _ven_bonds: usize,
-) -> Option<String> {
-    // Kalippaa typically has 4 lines
-    if lines.len() != 4 {
-        return None;
-    }
-
-    // Kalippaa has more flexible foot counts than Venkalippaa
-    // Common patterns: 4-3-4-3, 4-4-3-4, etc.
-    let mut valid_pattern = false;
-    let foot_counts: Vec<usize> = lines.iter().map(|l| l.feet.len()).collect();
-
-    // Check common Kalippaa patterns
-    if foot_counts == [4, 3, 4, 3]
-        || foot_counts == [4, 4, 3, 4]
-        || foot_counts == [3, 4, 3, 4]
-        || foot_counts == [4, 3, 3, 4]
-    {
-        valid_pattern = true;
-    }
-
-    if !valid_pattern {
-        return None;
-    }
-
-    // Kalippaa requires good bonding but less strict than Asiriyappaa
-    let talai_check = total_bonds > 0 && kali_bonds > 0;
-
-    if talai_check {
-        Some("kali_ppA".to_string())
-    } else {
-        None
-    }
-}
 
 // ==================== FUTURE EXPANSION POINTS (already planned) ====================
 // - GetLetterCount() → letter-by-letter Tamil analysis
@@ -989,51 +147,6 @@ mod tests {
 
         assert_eq!(parsed["lines"].as_array().unwrap().len(), 2);
         assert!(parsed["metre_type"].as_str().unwrap().contains("வெண்பா"));
-    }
-
-    #[test]
-    fn test_parse_poem_kaliviruttam() {
-        let text = "பேணநோற் றதுமனைப் பிறவி பெண்மைபோல்\nநாணநோற் றுயர்ந்தது நங்கை தோன்றலான்\nமாணநோற் றீண்டிவள் இருந்த வாறெலாம்\nகாணநோற் றிலனவன் கமலக் கண்களால்";
-
-        let result = parse_poem(text);
-        let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-
-        assert_eq!(parsed["lines"].as_array().unwrap().len(), 4);
-        assert!(parsed["metre_type"]
-            .as_str()
-            .unwrap()
-            .contains("kaliviru_tta_m"));
-    }
-
-    #[test]
-    fn test_debug_venpaa_2line() {
-        let text = "முற்ற உணர்ந்தானை ஏத்தி மொழிகுவன்\nகுற்றமொன்று இல்லா அறம்";
-
-        let result = parse_poem(text);
-        let parsed: serde_json::Value = serde_json::from_str(&result).unwrap();
-
-        println!("=== 2-LINE VENPAA DEBUG ===");
-        println!("Text: {}", text);
-        println!("Lines: {}", parsed["lines"].as_array().unwrap().len());
-        println!("Metre type: {}", parsed["metre_type"].as_str().unwrap());
-
-        // Check line structure
-        for (i, line) in parsed["lines"].as_array().unwrap().iter().enumerate() {
-            let feet = line["feet"].as_array().unwrap();
-            println!("Line {}: {} feet", i + 1, feet.len());
-            for (j, foot) in feet.iter().enumerate() {
-                let syllables = foot["syllables"].as_array().unwrap();
-                println!(
-                    "  Foot {}: {} ({})",
-                    j + 1,
-                    foot["foot_type"],
-                    syllables.len()
-                );
-            }
-        }
-
-        assert_eq!(parsed["lines"].as_array().unwrap().len(), 2);
-        // This should be venpaa, not venpaavinam
     }
 
     #[test]
@@ -1621,13 +734,88 @@ mod tests {
 
     #[test]
     fn test_check_venpaa_multiline() {
-        let lines = vec![Line {
-            feet: vec![Foot {
-                syllables: vec![],
-                foot_type: "tEmA".to_string(),
-            }],
-            line_class: "kuRaLaTi".to_string(),
-        }];
+        let lines = vec![
+            Line {
+                feet: vec![
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "tEmA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "puLimA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "kUviLa_m".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "karuviLa_m".to_string(),
+                    },
+                ],
+                line_class: "kuRaLaTi".to_string(),
+            },
+            Line {
+                feet: vec![
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "tEmA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "puLimA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "kUviLa_m".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "karuviLa_m".to_string(),
+                    },
+                ],
+                line_class: "kuRaLaTi".to_string(),
+            },
+            Line {
+                feet: vec![
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "tEmA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "puLimA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "kUviLa_m".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "karuviLa_m".to_string(),
+                    },
+                ],
+                line_class: "kuRaLaTi".to_string(),
+            },
+            Line {
+                feet: vec![
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "tEmA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "puLimA".to_string(),
+                    },
+                    Foot {
+                        syllables: vec![],
+                        foot_type: "kUviLa_m".to_string(),
+                    },
+                ],
+                line_class: "kuRaLaTi".to_string(),
+            },
+        ];
         assert_eq!(
             check_venpaa_multiline(&lines),
             Some("வெண்பா (Venpaa)".to_string())
