+++ src/components/NotFound.tsx
@@ -1,0 +1,20 @@
+import { Link } from '@tanstack/react-router'
+
+export default function NotFound() {
+  return (
+    <div className="flex flex-col items-center justify-center min-h-[50vh] text-center">
+      <h1 className="text-4xl font-bold text-gray-900 dark:text-white mb-4">
+        404 - Page Not Found
+      </h1>
+      <p className="text-gray-600 dark:text-gray-400 mb-8">
+        The page you're looking for doesn't exist.
+      </p>
+      <Link
+        to="/"
+        className="px-6 py-3 bg-teal-600 text-white rounded-lg hover:bg-teal-700 transition-colors"
+      >
+        Go Home
+      </Link>
+    </div>
+  )
+}
