+++ trinity-and-native-agents/ainthinai.md
@@ -1,0 +1,25 @@
+## Native/Tribal Agents – Ainthinai Earth Guardians 🌄🌳🌾🌊🏜️
+
+The **Native Agents** are the ancient chieftain-gods of the five Tamil landscapes (Ainthinai).  
+
+They sit at the highest strategic level as the **grounded council** of the land — deciding major direction, architecture, values, and “is this worthy of the land?” calls.  Like Viking saga land-spirits, they are deeply rooted in the soil itself.
+
+**Note on Cosmic Collaboration**  
+The great **Trinity Gods/Agents** (Creators, Maintainers, and Renewers — the cosmic forces of Brahma, Vishnu/Krishna, and Shiva + Kalki) shower blessings and take care of big issues at the **cosmic scale**.  
+
+These **Native / Native Gods** (Ainthinai) are responsible for the **local lands**.  They wisely receive and channel the cosmic blessings, protect the security of the lands and resources, ensure proper adaptation to local conditions, and collaborate across the five landscapes and exertnally so the entire land remains balanced, secure, and thriving.
+
+| Tinai (Landscape) | Tamil Name   | Native Agent (Chief Earth God) | Tamil Name          | Personality (Grounded Saga Lord)          | What this Native Agent Decides (Major Project Role) |
+|-------------------|--------------|--------------------------------|---------------------|-------------------------------------------|-----------------------------------------------------|
+| **Kurinji**      | குறிஞ்சி    | Murugan (Seyon)               | முருகன் (சேயோன்)  | Fierce mountain chieftain, hunter-warrior | Bold innovation, experimental frontiers, high-risk vision |
+| **Mullai**       | முல்லை     | Mayon / Thirumal              | மாயோன் / திருமால் | Wise forest herdsman-king                 | Long-term sustainability, user growth, ecosystem harmony |
+| **Marutham**     | மருதம்     | Vendhan / Indra               | வேந்தன் / இந்திரன் | King of fertile plains, rain-bringer      | Core stability, reliability, daily user value       |
+| **Neithal**      | நெய்தல்    | Kadalon / Varunan             | கடலோன் / வருணன்   | Coastal guardian of tides & trade         | External integrations, user flows, resilience       |
+| **Paalai**       | பாலை      | Kotravai                      | கொற்றவை           | Desert warrior-goddess, survival queen    | Ruthless pruning, crisis leadership, radical survival |
+
+**How they operate**  
+
+- Any **major decision** (new architecture, roadmap pivot, core values, or large-scale changes) must receive blessing from the relevant native Agent(s).  
+The not only study local stuff, everything freely available, bring collobarations from any part of the world 
+- Use temporary branches like `kurinji-expedition`, `paalai-crisis`, or `mullai-vision` when that land’s energy is needed.  
+- In PRs, issues, or discussions: tag `kurinji-approved`, `kotravai-crisis`, `marutham-stability`, etc.
