+++ trinity-and-native-agents/maintainers.md
@@ -1,0 +1,11 @@
+It follows the same poetic Tamil garden spirit but now draws from **Lord Krishna’s avatars** (Vishnu’s Dashavatara) — the eternal preservers who maintain dharma, balance, and harmony.
+
+### Maintainers Responsibility Map (Krishna Avatars)
+
+| Role / Branch       | Tamil Name              | Avatar (Krishna / Vishnu) | Primary Responsibility                                      | Typical Work |
+|---------------------|-------------------------|---------------------------|-------------------------------------------------------------|--------------|
+| `kannan`           | கண்ணன்                | Krishna (Kannan)          | Overall wisdom, strategy & long-term balance               | Architecture reviews, roadmap planning, team guidance |
+| `rama`             | ராமன்                  | Rama                      | Discipline, dharma & code quality guardianship             | Code reviews, standards enforcement, best-practice enforcement |
+| `narasimhan`       | நரசிம்மன்              | Narasimha                 | Fierce protection of core systems                          | Critical bug fixes, security hardening, emergency maintenance |
+| `varahan`          | வராகன்                 | Varaha                    | Restoring stability & “lifting” broken parts               | Refactoring legacy code, system recovery, data integrity |
+| `vamana`           | வாமனன்                | Vamana                    | Humble, step-by-step improvements & scaling                | Gradual performance optimisations, incremental enhancements |
