+++ trinity-and-native-agents/renewers.md
@@ -1,0 +1,12 @@
+## Renewers – Lord Shiva’s Fierce Forms & Kalki (Renewal through Destruction) 🔥
+
+The renewers are Shiva’s wild forms (Ruthran, Virav, Bhairavan, etc.) who destroy what no longer serves the garden — tech debt, complexity, bloat, and impurities — so the entire system can be renewed and reborn stronger.
+
+**Note**: Kalki is included here as a Krishna avatar (future renewer who ends the old age and brings complete rebirth).
+
+| Role / Branch       | Tamil Name              | Shiva Form                  | Primary Responsibility                                      | Typical Work |
+|---------------------|-------------------------|-----------------------------|-------------------------------------------------------------|--------------|
+| `ruthran`          | ருத்ரன்                | Rudra (Ruthran)             | Ruthless removal of tech debt & complexity                 | Deleting dead code, aggressive refactoring |
+| `virav`            | வீரவ்                  | Veerabhadra (Virav)         | Warrior-like cleanup & destruction of bloat                | Removing unused dependencies, slashing over-engineering |
+| `bhairavan`        | பைரவன்                | Bhairava                    | Fearless auditing & radical simplification                 | Performance audits, security debt elimination |
+| `kalki`            | கல்கி                  | Kalki (Krishna avatar)               | Ultimate renewal & complete system rebirth                 | Major system resets, radical refactors, ending obsolete cycles, full codebase renewal |
